refs#59 , craete a [chain] directory in /pages for chains in the future
