+++ packages/next/pages/[chain]/detail.js
@@ -9,7 +9,7 @@ import { timelineData } from "utils/data";
 import Comment from "components/comment";
 import { commentData } from "utils/data";
 import Position from "components/position";
-import {withLoginUser, withLoginUserRedux} from "../lib";
+import { withLoginUser, withLoginUserRedux } from "../../lib";
 
 const Wrapper = styled.div`
   > :not(:first-child) {
@@ -17,9 +17,7 @@ const Wrapper = styled.div`
   }
 `;
 
-export default withLoginUserRedux(({
-  loginUser,
-}) => {
+export default withLoginUserRedux(({ loginUser }) => {
   return (
     <Layout user={loginUser} right={<Position />}>
       <Wrapper>
@@ -34,7 +32,6 @@ export default withLoginUserRedux(({
 
 export const getServerSideProps = withLoginUser(async (context) => {
   return {
-    props: {
-    },
+    props: {},
   };
 });
