+++ packages/next/pages/[chain]/index.js
@@ -1,15 +1,13 @@
-import { discussionData } from "utils/data";
-import List from "components/list";
 import Layout from "components/layout";
+
+import Overview from "components/overview";
 import Menu from "components/menu";
 import Trends from "components/trends";
 import Footer from "components/footer";
 import { mainMenu } from "utils/constants";
-import {withLoginUser, withLoginUserRedux} from "../lib";
+import { withLoginUser, withLoginUserRedux } from "../../lib";
 
-export default withLoginUserRedux(({
-  loginUser,
-}) => {
+export default withLoginUserRedux(({ loginUser }) => {
   return (
     <Layout
       user={loginUser}
@@ -21,18 +19,13 @@ export default withLoginUserRedux(({
         </>
       }
     >
-      <List
-        category={discussionData.category}
-        items={discussionData.items}
-        pagination
-      />
+      <Overview />
     </Layout>
   );
 });
 
 export const getServerSideProps = withLoginUser(async (context) => {
   return {
-    props: {
-    },
+    props: {},
   };
 });
