+++ packages/next/pages/index.js
@@ -1,34 +1,19 @@
-import Layout from "components/layout";
+import { useEffect } from "react";
+import { useRouter } from "next/router";
 
-import Overview from "components/overview";
-import Menu from "components/menu";
-import Trends from "components/trends";
-import Footer from "components/footer";
-import { mainMenu } from "utils/constants";
-import {withLoginUser, withLoginUserRedux} from "../lib";
+export default function Root() {
+  const router = useRouter();
 
-export default withLoginUserRedux(({
-  loginUser,
-}) => {
-  return (
-    <Layout
-      user={loginUser}
-      left={<Menu menu={mainMenu} />}
-      right={
-        <>
-          <Trends />
-          <Footer />
-        </>
-      }
-    >
-      <Overview />
-    </Layout>
-  );
-});
+  useEffect(() => {
+    const chain = localStorage.getItem("chain") || "karura";
+    router.push(`/${chain}`);
+  }, [router]);
 
-export const getServerSideProps = withLoginUser(async (context) => {
+  return null;
+}
+
+export async function getServerSideProps(context) {
   return {
-    props: {
-    },
+    props: {},
   };
-});
+}
